feature(integration): add eye tracking in one file, need fix
