+++ eye-chat/eye-chat.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -15,7 +15,7 @@
       "location" : "https://github.com/apfdamascena/ios-eye-tracking",
       "state" : {
         "branch" : "master",
-        "revision" : "cf6e202b6dec9293ff32e519852c120b69e88389"
+        "revision" : "b8231d506a01fddf5f30dcbb9fabc149a14e18da"
       }
     }
   ],
