+++ eye-chat/eye-chat.xcodeproj/xcuserdata/alexdamascena.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -196,7 +196,7 @@
 		<key>eye-chat.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>1</integer>
+			<integer>0</integer>
 		</dict>
 	</dict>
 </dict>
