--- eye-chat/eye-chat/EyeTracking/Alias.swift
@@ -1,11 +1,0 @@
-//
-//  Alias.swift
-//  eye-chat
-//
-//  Created by alexdamascena on 03/04/23.
-//
-
-import Foundation
-
-
-typealias Function = (() -> ())
