+++ eye-chat/eye-chat/EyeTracking/AppState.swift
@@ -1,0 +1,8 @@
+//
+//  AppState.swift
+//  eye-chat
+//
+//  Created by alexdamascena on 05/04/23.
+//
+
+import Foundation
