--- eye-chat/eye-chat/EyeTracking/AttributesTrigger.swift
@@ -1,19 +1,0 @@
-//
-//  AttributesTrigger.swift
-//  eye-chat
-//
-//  Created by alexdamascena on 03/04/23.
-//
-
-import Foundation
-
-
-struct AttributesTrigger {
-    
-    var axisXFirstSectionTrigger: [Function]?
-    var axisXSecondSectionTrigger: [Function]?
-    var upTrigger: Function?
-    var downTrigger: Function?
-    var goBack: Function?
-    
-}
