+++ eye-chat/eye-chat/EyeTracking/EyeTrackingView.swift
@@ -2,17 +2,19 @@
 //  EyeTrackingView.swift
 //  eye-chat
 //
-//  Created by alexdamascena on 03/04/23.
+//  Created by alexdamascena on 05/04/23.
 //
 
 import SwiftUI
+import EyeTracking
+import ARKit
 
 struct EyeTrackingView: UIViewControllerRepresentable {
     
     typealias UIViewType = EyeTrackingViewController
     
     @ObservedObject var viewModel: EyeTrackingViewModel
-    @Binding var endEyeTracking: Bool
+    @Binding var endEyeTracking: Int
     
     func makeUIViewController(context: Context) -> EyeTrackingViewController {
         let controller = EyeTrackingViewController()
@@ -21,8 +23,10 @@ struct EyeTrackingView: UIViewControllerRepresentable {
     }
     
     func updateUIViewController(_ uiViewController: EyeTrackingViewController, context: Context) {
-        if endEyeTracking {
-            uiViewController.eyeTracking.hidePointer()
+        
+        if endEyeTracking == -1 {
+            AppState.shared.gameID = UUID()
         }
+ 
     }
 }
