+++ eye-chat/eye-chat/UI/Components/Teste.swift
@@ -1,0 +1,20 @@
+//
+//  Teste.swift
+//  eye-chat
+//
+//  Created by alexdamascena on 05/04/23.
+//
+
+import SwiftUI
+
+struct Teste: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct Teste_Previews: PreviewProvider {
+    static var previews: some View {
+        Teste()
+    }
+}
