+++ eye-chat/eye-chat/UI/Modules/Opa/Opa.swift
@@ -1,0 +1,13 @@
+//
+//  Opa.swift
+//  eye-chat
+//
+//  Created by alexdamascena on 05/04/23.
+//  
+//
+
+import SwiftUI
+
+struct Opa {
+    
+}
