+++ eye-chat/eye-chat/UI/Modules/Opa/OpaViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  OpaViewModel.swift
+//  eye-chat
+//
+//  Created by alexdamascena on 05/04/23.
+//  
+//
+
+import SwiftUI
+
+// MARK: Variables to watch mark as Published. ViewModel also calls API/Core Data
+
+final class OpaViewModel: ObservableObject {
+    @Published var example: String?
+}
