+++ eye-chat/eye-chat/eye_chatApp.swift
@@ -9,12 +9,32 @@ import SwiftUI
 
 @main
 struct eye_chatApp: App {
+    
     let persistenceController = PersistenceController.shared
+    @ObservedObject var coordinator = ViewCordinator()
+    
+    @StateObject var appState = AppState.shared
 
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            if #available(iOS 16.0, *){
+                NavigationStack(path: $coordinator.path){
+                    HomeV2View()
+                    .id(appState.gameID)
+                    .navigationDestination(for: RouteScreen.self){ destination in
+                        switch destination {
+                        case .main:
+                            ThirdView()
+                        case .create:
+                            ThirdView()
+                        }
+                    }
+                }
                 .environment(\.managedObjectContext, persistenceController.container.viewContext)
+                .environmentObject(coordinator)
+            }
+
+             
         }
     }
 }
